Update Assemble.py
Here's an idea for implementing Assemble more generally. Since all subclasses need to do is configure the assembler and set the targets, we can generalize so that the only thing the user needs to override is an "aim" method that returns target vertices, and a "config" method that sets tunables (i.e. stepRate) and also declares any parameters used in 'aim'.

Feedback appreciated - I'm not sure if this is the optimal structure. For example, we don't technically need two methods; we could just set all the assembler parameters and aim parameters in "aim" and override that. Or, we technically can just override "dotask"  instead of defining a new function 'aim'. Thoughts...?