Update QTaskmanager.py
Tentative background task structure. At all times, self.task is the task at the front of the queue, and background tasks never enter the queue. (Instead background tasks are kept in a list, for reference)
Registering:    Background tasks are registered when blocking=False, and are connected without going through the queue. 
Deregistering:  Background tasks must always pass a task (itself) to deactivateTask. If task=None, then the task is assumed to be a blocking task and is taken from the queue (i.e. self.task). 

Review before pulling - code not yet tested.