Create QSubtaskManager.py
Idea for structuring subtasks. If, for instance, vision.py is a QTask that detects particles, and another task is designed to read vision signals and plot positions on a graph, then how will the subtask connect to the Vision task's output? One way would be to make a new task manager for vision's subtasks, but with vision's output as a source. (I.e. vision has its own signal sigNewFrame(Trajectories) that passes trajectories to its subtasks). 

To make this work, the task manager needs to pass subtasks to the appropriate subtask manager. We can do this by organizing like-subtasks into the same subfolder in 'lib'.
The main task manager also needs to properly shut down the subtask manager when the "source" task (i.e. vision) ends (i.e. recieves sigDone() from the source).