Connect structureChanged to cgh.computeTrap(self)
cgh was given a method computeTrap(trap) which updates the argument (trap)'s hologram. So, structure changes signal the CGH to update the individual trap, but not the entire hologram. computeTrap can be moved into the QTrap class later, if desired. We can also probably remove all of the refresh() framework too