Reimplement some TrapMove elements. I can't spell 'parameterize'
New: Read all kwargs into object and pass in parameterize properties  on __init__; so subclasses only have to override parameterize. This will allow parameterize to be called on initialize instead of __init__ if desired

Come back to this and see if it's smart or just overly complicated; and take it out if needed. 

Reimplement TrapMove traps setter. Account for trap, list of traps, or QTrapGroup input types. **** Is 'select' outdated?? Also, how do you update a single trap, without access to QTrap class (and enum type state?)****

i THINK the refresh() functions in process() are outdated, but still need to check.

Allow trajectory setter to be a dict