Replace Trajectory objects with lists
The old code stored Trajectory objects, with Trajectory.data a list of vertices. Now, we simply use the vertices. 
stepSize is declared in trapMove and not used here, so we remove the setter. The getter still overrides the parent class, though, so we keep it. 
stepSize is also not used in parametrize

Replace tmax with nframes, and remove t, tf