Add various tasks to demonstrate TaskData
setTaskData is the highlight! This is a Task that's only function is to receive input, and save that input as taskData for the next task. Calling this task from the menu prompts the user to enter the task data via QInputBox.
An example of this is in TaskDataExample3, where the user can type 'xc' and 'yc' and their corresponding values to set the center for Encircle, instead of going into the Encircle file and changing it manually (which requires pyfab to restart). 
 
GlobalCounter is useful for testing background tasks, and timing # of frames between events (such as registering a task and that task finishing __init__, or timing how many frames initialize() takes to finish). Also used to demonstrate passing task data.


Note: In general, the screen and the terminal output freeze while init() is running, which is noticable for tasks which take longer than a frame to start. This is why assemble glitches the UI. One possible workaround could be to try running Assemble as a background task that calculates trajectories and then passes them to a Move task - though this needs more thought.